address the comments in the pr
