+++ mtags-shared/src/main/scala/scala/meta/internal/pc/CompilerJobQueue.scala
@@ -1,6 +1,10 @@
 package scala.meta.internal.pc
 
-import java.util.concurrent._
+import java.util.concurrent.CancellationException
+import java.util.concurrent.CompletableFuture
+import java.util.concurrent.PriorityBlockingQueue
+import java.util.concurrent.ThreadPoolExecutor
+import java.util.concurrent.TimeUnit
 import java.util.concurrent.atomic.AtomicReference
 import java.{util => ju}
