+++ tests/unit/src/test/scala/tests/codeactions/ExtractValueLspSuite.scala
@@ -147,7 +147,7 @@ class ExtractValueLspSuite
     s"""|object Main {
         |  def method2(i: Int) : Int = ???
         |  def method1(s: String): Unit = {
-        |    println("Hello!");${' '}
+        |    println("Hello!"); 
         |    val newValue = 1 + 2
         |    method2(1 + method2(newValue))
         |  }
